Implement collapsing sidebar with a button for narrow view width
Making use of absolute positioning to "stick" the sidebar to the right

Using width 0 and 80% to "collapse" and "show" sidebar respectively
Transitions to make the collapsing and showing seem "dynamic"

Making the toggle button a client component to use state to determine
when to show/hide the sidebar as well as which icons to use.

Applying some filters to the main article element to make it appear as
if it's "fading off". Created a CSS class for this, and toggling this
class on the main article element.

Use lg breakpoint for the width setting on the main article and sidebar.
Allows the main article to take the full width, as it should, when it's
a narrow view e.g. on mobile.