Abstract avatar into its component while removing the UserIcon component
The avatar could benefit being its own component as it's used in several
other components, and currently, abstraction for this seems to be fine.

The previous UserIcon component does a bit more than just icons
delivering other username / display name things. It seemed to be
specific to the sidebar. So moved the username / display name code to
the sidebar while using the new Avatar component, and removed this
UserIcon component.